# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School API",
        "description": "API for managing school data",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/api/v1/fees/structures": {
            "get": {
                "tags": [
                    "school-v2.4",
                    "school-v2.0"
                ],
                "summary": "List fee structures",
                "operationId": "f0cde3b78c5ca8834280d9a6d3e089f1",
                "responses": {
                    "200": {
                        "description": "Fee structures returned"
                    }
                }
            },
            "post": {
                "tags": [
                    "school-v2.4",
                    "school-v2.0"
                ],
                "summary": "Create fee structure",
                "operationId": "1b5e9cc4f0bff2114496186f1de7eb82",
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error or duplicate"
                    }
                }
            }
        },
        "/api/v1/fees/structures/{feeStructure}": {
            "get": {
                "tags": [
                    "school-v2.4",
                    "school-v2.0"
                ],
                "summary": "Get fee structure",
                "operationId": "4e0ad488bc3c5915149362dead358126",
                "parameters": [
                    {
                        "name": "feeStructure",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fee structure returned"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "school-v2.4",
                    "school-v2.0"
                ],
                "summary": "Update fee structure",
                "operationId": "d06c3ed882211ff96b6d6b22e5fd6d1f",
                "parameters": [
                    {
                        "name": "feeStructure",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "school-v2.4",
                    "school-v2.0"
                ],
                "summary": "Delete fee structure",
                "operationId": "f3a6257786917f17217a1eb63aa9e787",
                "parameters": [
                    {
                        "name": "feeStructure",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/fees/structures/copy": {
            "post": {
                "tags": [
                    "school-v2.4",
                    "school-v2.0"
                ],
                "summary": "Copy fee structures between sessions/terms",
                "operationId": "55f50832268f7702ebeb21425583c08c",
                "responses": {
                    "201": {
                        "description": "Structures copied"
                    },
                    "404": {
                        "description": "Source not found"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}